added demo script
