+++ .gitignore
@@ -7,3 +7,4 @@
 !*.c
 !*.h
 !*.py
+!*demo
